codex32/requirements.txt
# Core Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Async & Concurrency
asyncio-contextmanager==1.0.1
# NOTE: dependency-injector is optional for this repo (we use lightweight file-based DI).
# It currently fails to build on CPython 3.14 due to C-extension incompatibility.

# Data & Configuration
# NOTE: CPython 3.14 currently requires Rust builds for pydantic-core (Pydantic v2).
# To keep installs "free" and frictionless, we use the pure-Python Pydantic v1 line.
pydantic==1.10.22
sqlalchemy==2.0.23
asyncpg==0.29.0 ; python_version < "3.13"
psycopg[binary,pool]==3.3.2
alembic==1.12.1

# Redis & Caching
redis[hiredis]==5.0.1
aioredis==2.0.1

# Database Drivers
# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
cryptography==41.0.7

# HTTP Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# System & Process Management
psutil==5.9.6
PyYAML==6.0.2

# WebSocket & Real-time Communication
websockets==12.0
python-socketio==5.10.0
python-engineio==4.8.0

# Speech-to-Text & Audio Processing
vosk==0.3.44 ; python_version < "3.13"
SpeechRecognition==3.10.0 ; python_version < "3.13"
pydub==0.25.1 ; python_version < "3.13"
soundfile==0.12.1 ; python_version < "3.13"
librosa==0.10.0 ; python_version < "3.13"

# Optional: Cloud Speech Services
google-cloud-speech==2.22.0 ; python_version < "3.13"
azure-cognitiveservices-speech==1.47.0 ; python_version < "3.13"

# Conversational AI (Optional)
# rasa==3.5.0  # Uncomment if using Rasa

# Data Science & ML
scikit-learn==1.3.2 ; python_version < "3.13"
joblib==1.3.2 ; python_version < "3.13"
pandas==2.1.3 ; python_version < "3.13"
numpy==1.26.2 ; python_version < "3.13"

# Monitoring & Metrics
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Logging & Structured Logs
python-json-logger==2.0.7
structlog==23.2.0

# Environment & Dotenv
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
# httpx does not provide a 'testing' extra; keep base package pinned above.
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Utilities
python-magic==0.4.27
tenacity==8.2.3
